// angularProject/src/app/api/model/dropDown.ts
/**
 * TrueValue
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DropDown { 
    id?: number;
    fieldName?: string | null;
    fieldType?: string | null;
    options?: string | null;
    pageName?: string | null;
    custField1?: string | null;
    custField2?: string | null;
    custField3?: string | null;
}